+++ src/Platform/CAN/TPMessage.cpp
@@ -2,7 +2,7 @@
 
 using IdInfo = CAN::TPMessage::IdInfo;
 
-IdInfo CAN::TPMessage::decodeId(uint16_t canID) {
+IdInfo CAN::TPMessage::decodeId(uint32_t canID) {
     idInfo.sourceAddress = (canID >> 4) & 0b111;
     idInfo.destinationAddress = (canID >> 1) & 0b111;
     idInfo.isMulticast = canID & 0b1;
@@ -10,7 +10,7 @@ IdInfo CAN::TPMessage::decodeId(uint16_t canID) {
     return idInfo;
 }
 
-uint16_t CAN::TPMessage::encodeId() const {
+uint32_t CAN::TPMessage::encodeId() const {
     uint16_t id = 0b0111 << 7;
 
     id |= idInfo.sourceAddress << 4;
